refactor(apps): extract onsite ReviewModal from review.tsx into testable OnsiteReviewModal + add missing mod-review UI tests
Extract the on-site (App Block) `ReviewModal` (and its manifest/diff/preview
sub-panels + request types + byte-formatters) out of the monolithic
`src/pages/apps/review.tsx` into a new `src/components/Apps/OnsiteReviewModal.tsx`,
mirroring the #3154 extraction of the diff panels. `review.tsx` exports
`getServerSideProps`, so the modal could not be imported into a browser (jsdom)
test while inline; the new module is server-graph-free (its only `~/server`
imports are a zod schema + two constant tables) so it is now unit-testable.
`review.tsx` imports and mounts it identically — zero behaviour change.

Add the missing moderator-review UI coverage:
- OnsiteReviewModal.browser.test.tsx: onsite Approve fires blocks.approveRequest
  with the request id; onsite reject reason-gate (disabled < 3 chars) + Reject
  fires blocks.rejectRequest with {publishRequestId, rejectionReason}; read-only
  (approved) posture shows the mod feedback and no actions.
- OffsiteReviewQueue.browser.test.tsx: assert offsite Reject actually FIRES
  rejectExternalRequest (previously only the gate was asserted).
- AppListingsModerationTable.browser.test.tsx: assert Relist on a removed row
  FIRES relistListing (the sole recovery path for a mistaken takedown; button
  presence was asserted but never clicked/fired).

Dark / mod-gated (Flipt appListings/appBlocks). No migration, no behaviour change.
