+++ src/components/Apps/OffsiteReviewQueue.browser.test.tsx
@@ -30,6 +30,7 @@ const OFFSITE_ROW = {
 const mocks = vi.hoisted(() => ({
   invalidate: vi.fn().mockResolvedValue(undefined),
   approveMutate: vi.fn(),
+  rejectMutate: vi.fn(),
 }));
 
 vi.mock('~/providers/FeatureFlagsProvider', () => ({
@@ -83,7 +84,13 @@ vi.mock('~/utils/trpc', () => {
             isPending: false,
           }),
         },
-        rejectExternalRequest: { useMutation: mutation },
+        rejectExternalRequest: {
+          useMutation: () => ({
+            mutate: mocks.rejectMutate,
+            mutateAsync: vi.fn(),
+            isPending: false,
+          }),
+        },
       },
     },
   };
@@ -94,6 +101,7 @@ const { OffsiteReviewQueue } = await import('./OffsiteReviewQueue');
 beforeEach(() => {
   mocks.invalidate.mockClear();
   mocks.approveMutate.mockClear();
+  mocks.rejectMutate.mockClear();
 });
 
 describe('OffsiteReviewQueue — kind-aware review row', () => {
@@ -172,6 +180,24 @@ describe('OffsiteReviewModal — approve-notes gating, friendly date, field labe
     // Whitespace-only padding does NOT satisfy the gate (trimmed length counts).
     await page.getByTestId('apps-offsite-reject-reason').fill('  a  ');
     await expect.element(confirm).toBeDisabled();
+  });
+
+  // Beyond the GATE (disabled→enabled), lock that a valid reject actually FIRES
+  // the reject mutation with the trimmed reason + the request id — no prior test
+  // asserted the fired offsite reject mutation (only the gate).
+  test('Reject with a ≥min-length reason FIRES rejectExternalRequest with {publishRequestId, rejectionReason}', async () => {
+    renderWithProviders(<OffsiteReviewQueue />);
+    await page.getByRole('button', { name: 'Review' }).click();
+    await page.getByTestId('apps-offsite-reject-open').click();
+
+    await page.getByTestId('apps-offsite-reject-reason').fill('needs a real reason');
+    const confirm = page.getByTestId('apps-offsite-reject-confirm');
+    await expect.element(confirm).toBeEnabled();
+    await confirm.click();
+    expect(mocks.rejectMutate).toHaveBeenCalledWith({
+      publishRequestId: 'req-1',
+      rejectionReason: 'needs a real reason',
+    });
   });
 
   // The disabled-reason Tooltip wraps the disabled Button in a Box so it still
