+++ src/components/Apps/OnsiteReviewModal.tsx
@@ -1,0 +1,1439 @@
+import {
+  Accordion,
+  Alert,
+  Badge,
+  Box,
+  Button,
+  Card,
+  Code,
+  Group,
+  Modal,
+  NumberInput,
+  ScrollArea,
+  Select,
+  SimpleGrid,
+  Stack,
+  Switch,
+  Text,
+  Textarea,
+  Tooltip,
+} from '@mantine/core';
+import {
+  IconAdjustmentsAlt,
+  IconAlertTriangle,
+  IconCheck,
+  IconCode,
+  IconExternalLink,
+  IconKey,
+  IconLayoutGrid,
+  IconShieldLock,
+  IconWindow,
+  IconX,
+} from '@tabler/icons-react';
+import { useMemo, useRef, useState } from 'react';
+import { pickReviewIframeSrc } from '~/components/Apps/reviewIframeSrc';
+import {
+  FileDiffEntry,
+  FileListPreview,
+  ManifestDiffPreview,
+  type FileLineDiff,
+} from '~/components/Apps/reviewDiffPanels';
+import { useFeatureFlags } from '~/providers/FeatureFlagsProvider';
+import { OFFSITE_MOD_REASON_MIN } from '~/server/schema/blocks/offsite-moderation.schema';
+import {
+  MARKETPLACE_CATEGORIES,
+  MARKETPLACE_CATEGORY_LABELS,
+  type MarketplaceCategory,
+} from '~/server/services/blocks/marketplace-categories.constants';
+import {
+  SCOPE_DESCRIPTIONS,
+  SLOT_DESCRIPTIONS,
+} from '~/server/services/blocks/scope-descriptions.constants';
+import { showErrorNotification, showSuccessNotification } from '~/utils/notifications';
+import { trpc } from '~/utils/trpc';
+
+/**
+ * On-site (App Block) moderator review modal — extracted from `src/pages/apps/review.tsx`
+ * so it (and its manifest/diff/preview sub-panels) can be imported into a browser
+ * (jsdom) test WITHOUT pulling the page's `getServerSideProps`/`createServerSideProps`
+ * tRPC-server graph. This mirrors the #3154 extraction of the diff panels to
+ * `reviewDiffPanels.tsx` for the same reason. The sibling off-site modal is
+ * `OffsiteReviewModal` in `OffsiteReviewQueue.tsx`.
+ *
+ * Everything here is server-graph-free: the only `~/server/...` imports are pure
+ * constant tables (`scope-descriptions.constants`, `marketplace-categories.constants`)
+ * and a zod-only schema (`offsite-moderation.schema`) — none touch Prisma/tRPC server
+ * code. `review.tsx` imports these types/helpers + `OnsiteReviewModal` back and mounts
+ * it exactly as before (zero behaviour change).
+ */
+
+export type ManifestDiffSummary =
+  | { kind: 'first-version'; fields: string[] }
+  | {
+      kind: 'update';
+      added: string[];
+      removed: string[];
+      changed: Array<{ field: string; from: unknown; to: unknown }>;
+    };
+
+export type FileSummary = {
+  files: Array<{ path: string; sha256: string; sizeBytes: number }>;
+  added: string[];
+  removed: string[];
+  changed: string[];
+};
+
+export type UserProfile = { id: number; username: string | null; image: string | null };
+
+export type ReviewedRequestCommon = {
+  id: string;
+  appBlockId: string | null;
+  slug: string;
+  version: string;
+  submittedAt: string | Date;
+  bundleSizeBytes: string;
+  bundleSha256: string;
+  manifest: unknown;
+  fileSummary: unknown;
+  manifestDiffSummary: unknown;
+  reviewRepoUrl: string;
+  // PUSH-ORIGINATED requests (git-push, empty bundle) have no review-org
+  // snapshot — this links the mod to the canonical repo at the exact pushed sha.
+  pushCommitUrl?: string | null;
+  submittedBy: UserProfile;
+};
+
+export type PendingRequest = ReviewedRequestCommon;
+
+export type ApprovedRequest = ReviewedRequestCommon & {
+  reviewedAt: string | Date | null;
+  approvalNotes: string | null;
+  reviewedBy: UserProfile | null;
+};
+
+export type RejectedRequest = ReviewedRequestCommon & {
+  reviewedAt: string | Date | null;
+  rejectionReason: string | null;
+  reviewedBy: UserProfile | null;
+};
+
+export type AnyRequest = PendingRequest | ApprovedRequest | RejectedRequest;
+
+/** The review-modal display mode. Structurally identical to the review page's
+ *  `TabValue` (they share the same literal union), so the page's `selected.mode`
+ *  is assignable to this without importing from the page module. */
+export type OnsiteReviewMode = 'pending' | 'approved' | 'rejected' | 'reports';
+
+export type OnsiteReviewSelection = { request: AnyRequest; mode: OnsiteReviewMode } | null;
+
+export function formatBytes(s: string): string {
+  const n = Number(s);
+  if (!Number.isFinite(n)) return s;
+  if (n < 1024) return `${n} B`;
+  if (n < 1024 * 1024) return `${(n / 1024).toFixed(1)} KiB`;
+  return `${(n / (1024 * 1024)).toFixed(2)} MiB`;
+}
+
+export function formatDate(d: string | Date | null | undefined): string {
+  if (!d) return '—';
+  const date = typeof d === 'string' ? new Date(d) : d;
+  return date.toLocaleString();
+}
+
+// ---------------------------------------------------------------------------
+// Review modal — pending requests get the interactive approve/reject UI;
+// history requests get a read-only view with the mod feedback surfaced
+// prominently.
+// ---------------------------------------------------------------------------
+
+export function OnsiteReviewModal({
+  selection,
+  onClose,
+}: {
+  selection: OnsiteReviewSelection;
+  onClose: () => void;
+}) {
+  const utils = trpc.useUtils();
+  const [approvalNotes, setApprovalNotes] = useState('');
+  const [rejectionReason, setRejectionReason] = useState('');
+  const [actionMode, setActionMode] = useState<'view' | 'reject'>('view');
+
+  const approveMut = trpc.blocks.approveRequest.useMutation({
+    onSuccess: async () => {
+      showSuccessNotification({
+        message: `Approved ${selection?.request.slug} v${selection?.request.version}. Build started.`,
+      });
+      await utils.blocks.listPendingRequests.invalidate();
+      await utils.blocks.listApprovedRequests.invalidate();
+      onClose();
+    },
+    onError: (e) => {
+      showErrorNotification({
+        title: 'Approve failed',
+        error: new Error(e.message),
+      });
+    },
+  });
+  const rejectMut = trpc.blocks.rejectRequest.useMutation({
+    onSuccess: async () => {
+      showSuccessNotification({
+        message: `Rejected ${selection?.request.slug} v${selection?.request.version}.`,
+      });
+      await utils.blocks.listPendingRequests.invalidate();
+      await utils.blocks.listRejectedRequests.invalidate();
+      onClose();
+    },
+    onError: (e) => {
+      showErrorNotification({
+        title: 'Reject failed',
+        error: new Error(e.message),
+      });
+    },
+  });
+
+  if (!selection) return null;
+  const { request, mode } = selection;
+  const readOnly = mode !== 'pending';
+
+  const manifest = request.manifest as Record<string, unknown>;
+  const fs = (request.fileSummary ?? {}) as FileSummary;
+  const mds = (request.manifestDiffSummary ?? {}) as ManifestDiffSummary;
+  const busy = approveMut.isPending || rejectMut.isPending;
+
+  const approved = mode === 'approved' ? (request as ApprovedRequest) : null;
+  const rejected = mode === 'rejected' ? (request as RejectedRequest) : null;
+
+  return (
+    <Modal
+      opened={!!selection}
+      onClose={() => {
+        if (busy) return;
+        setApprovalNotes('');
+        setRejectionReason('');
+        setActionMode('view');
+        onClose();
+      }}
+      title={
+        <Group gap={6}>
+          <Text fw={600}>{request.slug}</Text>
+          <Code>{request.version}</Code>
+          {mds.kind === 'first-version' && (
+            <Badge color="violet" size="sm">
+              first version
+            </Badge>
+          )}
+          {mode === 'approved' && (
+            <Badge color="green" size="sm">
+              approved
+            </Badge>
+          )}
+          {mode === 'rejected' && (
+            <Badge color="red" size="sm">
+              rejected
+            </Badge>
+          )}
+        </Group>
+      }
+      size="xl"
+      centered
+    >
+      <Stack gap="md">
+        <Group gap="xs" align="flex-start">
+          <Text size="xs" c="dimmed">
+            Submitter:
+          </Text>
+          <Text size="xs">
+            {request.submittedBy.username ?? `#${request.submittedBy.id}`}
+          </Text>
+          <Text size="xs" c="dimmed">
+            ·
+          </Text>
+          <Text size="xs" c="dimmed">
+            {formatDate(request.submittedAt)}
+          </Text>
+          <Text size="xs" c="dimmed">
+            · {formatBytes(request.bundleSizeBytes)} · sha256:
+          </Text>
+          <Code style={{ fontSize: 10 }}>{request.bundleSha256.slice(0, 12)}…</Code>
+        </Group>
+
+        {(approved || rejected) && (
+          <Alert
+            color={approved ? 'green' : 'red'}
+            variant="light"
+            icon={approved ? <IconCheck size={16} /> : <IconX size={16} />}
+            title={
+              <Group gap={6}>
+                <Text size="sm" fw={600}>
+                  {approved ? 'Approved by' : 'Rejected by'}{' '}
+                  {(approved ?? rejected)?.reviewedBy
+                    ? (approved ?? rejected)!.reviewedBy!.username
+                      ? `@${(approved ?? rejected)!.reviewedBy!.username}`
+                      : `#${(approved ?? rejected)!.reviewedBy!.id}`
+                    : 'unknown'}
+                </Text>
+                <Text size="xs" c="dimmed">
+                  · {formatDate((approved ?? rejected)!.reviewedAt)}
+                </Text>
+              </Group>
+            }
+          >
+            {approved && approved.approvalNotes && (
+              <Stack gap={2}>
+                <Text size="xs" c="dimmed">
+                  Approval notes
+                </Text>
+                <Text size="sm" style={{ whiteSpace: 'pre-wrap', wordBreak: 'break-word' }}>
+                  {approved.approvalNotes}
+                </Text>
+              </Stack>
+            )}
+            {approved && !approved.approvalNotes && (
+              <Text size="xs" c="dimmed" fs="italic">
+                No approval notes were recorded.
+              </Text>
+            )}
+            {rejected && rejected.rejectionReason && (
+              <Stack gap={2}>
+                <Text size="xs" c="dimmed">
+                  Rejection reason
+                </Text>
+                <Text size="sm" style={{ whiteSpace: 'pre-wrap', wordBreak: 'break-word' }}>
+                  {rejected.rejectionReason}
+                </Text>
+              </Stack>
+            )}
+          </Alert>
+        )}
+
+        <Button
+          component="a"
+          href={request.pushCommitUrl ?? request.reviewRepoUrl}
+          target="_blank"
+          rel="noopener"
+          variant="default"
+          leftSection={<IconCode size={14} />}
+          rightSection={<IconExternalLink size={12} />}
+        >
+          {request.pushCommitUrl
+            ? 'View code in Forgejo (git push) ↗ commit'
+            : 'View code in Forgejo'}
+        </Button>
+
+        {/* MOD REVIEW SANDBOX (#2831) — run the PENDING version in a temporary,
+            mod-gated preview before approving. Pending requests only; dark
+            unless the mod-only review-sandbox flag is enabled. */}
+        {mode === 'pending' && <ReviewPreviewPanel publishRequestId={request.id} slug={request.slug} />}
+
+        {/* F-E E5 — publisher screenshot gallery review. Publisher-supplied
+            images are an abuse vector → the mod sees them (here, derived from
+            the submitted bundle) before approving. Renders for every mode so an
+            approved app's screenshots can be re-checked too. */}
+        <ScreenshotsReviewPanel publishRequestId={request.id} />
+
+        {/* F-E E4 curation — marketplace metadata (category / featured / order).
+            Only for an APPROVED request that has a linked app_block: featuring
+            is approved-only, and the meta lives on the app_block row. */}
+        {mode === 'approved' && request.appBlockId && (
+          <CurationPanel key={request.appBlockId} appBlockId={request.appBlockId} />
+        )}
+
+        <Stack gap={4}>
+          <Text size="sm" fw={600}>
+            Files
+          </Text>
+          <Group gap={6}>
+            <Text size="sm">{fs.files?.length ?? 0} total</Text>
+            {(fs.added?.length ?? 0) > 0 && (
+              <Badge color="green" variant="light">
+                +{fs.added.length} added
+              </Badge>
+            )}
+            {(fs.changed?.length ?? 0) > 0 && (
+              <Badge color="yellow" variant="light">
+                ~{fs.changed.length} changed
+              </Badge>
+            )}
+            {(fs.removed?.length ?? 0) > 0 && (
+              <Badge color="red" variant="light">
+                −{fs.removed.length} removed
+              </Badge>
+            )}
+          </Group>
+          {mds.kind === 'update' && (
+            <FileListPreview added={fs.added} removed={fs.removed} changed={fs.changed} />
+          )}
+          {/* Line-level code diff — lazy (only fetched when the mod toggles it
+              open) so the modal stays light by default. Bounded server-side;
+              binary / oversized / huge-diff files fall back to the Forgejo link
+              above. */}
+          <CodeDiffPanel
+            publishRequestId={request.id}
+            forgejoUrl={request.pushCommitUrl ?? request.reviewRepoUrl}
+          />
+        </Stack>
+
+        <Stack gap={4}>
+          <Text size="sm" fw={600}>
+            Manifest diff
+          </Text>
+          {mds.kind === 'first-version' ? (
+            <Text size="xs" c="dimmed">
+              First version — full manifest below.
+            </Text>
+          ) : (
+            <ManifestDiffPreview diff={mds} />
+          )}
+        </Stack>
+
+        <Stack gap={4}>
+          <Text size="sm" fw={600}>
+            Manifest
+          </Text>
+          <ManifestView manifest={manifest} />
+        </Stack>
+
+        {readOnly ? null : actionMode === 'reject' ? (
+          (() => {
+            const reasonLen = rejectionReason.trim().length;
+            const reasonTooShort = reasonLen < OFFSITE_MOD_REASON_MIN;
+            const rejectDisabled = busy || reasonTooShort;
+            return (
+              <Stack gap="xs">
+                <Text size="sm" fw={600}>
+                  Rejection reason
+                </Text>
+                <Textarea
+                  autosize
+                  minRows={3}
+                  maxRows={10}
+                  placeholder={`Explain what needs to change before this can be approved (≥${OFFSITE_MOD_REASON_MIN} chars, shown to the dev).`}
+                  description={`${reasonLen}/${OFFSITE_MOD_REASON_MIN} characters minimum`}
+                  error={
+                    reasonTooShort && reasonLen > 0
+                      ? `Enter at least ${OFFSITE_MOD_REASON_MIN} characters.`
+                      : undefined
+                  }
+                  value={rejectionReason}
+                  onChange={(e) => setRejectionReason(e.currentTarget.value)}
+                  disabled={busy}
+                  data-testid="apps-review-reject-reason"
+                />
+                <Group justify="flex-end" gap="xs">
+                  <Button variant="default" onClick={() => setActionMode('view')} disabled={busy}>
+                    Cancel
+                  </Button>
+                  <Tooltip
+                    label={`Enter a reason — at least ${OFFSITE_MOD_REASON_MIN} characters.`}
+                    disabled={!reasonTooShort}
+                    withArrow
+                  >
+                    {/* A native disabled <button> fires no pointer events, so the
+                        Tooltip must attach to a wrapper to show in the exact state
+                        it explains (Mantine's documented disabled-target pattern). */}
+                    <Box>
+                      <Button
+                        color="red"
+                        leftSection={<IconX size={14} />}
+                        onClick={() =>
+                          rejectMut.mutate({
+                            publishRequestId: request.id,
+                            rejectionReason: rejectionReason.trim(),
+                          })
+                        }
+                        disabled={rejectDisabled}
+                        loading={rejectMut.isPending}
+                        data-testid="apps-review-reject-confirm"
+                      >
+                        Reject
+                      </Button>
+                    </Box>
+                  </Tooltip>
+                </Group>
+              </Stack>
+            );
+          })()
+        ) : (
+          <Stack gap="xs">
+            <Textarea
+              label="Approval notes (optional)"
+              autosize
+              minRows={2}
+              maxRows={6}
+              placeholder="Optional notes attached to the approval record."
+              value={approvalNotes}
+              onChange={(e) => setApprovalNotes(e.currentTarget.value)}
+              disabled={busy}
+            />
+            <Group justify="flex-end" gap="xs">
+              <Button
+                color="red"
+                variant="default"
+                leftSection={<IconX size={14} />}
+                onClick={() => setActionMode('reject')}
+                disabled={busy}
+              >
+                Reject…
+              </Button>
+              <Button
+                color="green"
+                leftSection={<IconCheck size={14} />}
+                onClick={() =>
+                  approveMut.mutate({
+                    publishRequestId: request.id,
+                    approvalNotes: approvalNotes.trim() || undefined,
+                  })
+                }
+                disabled={busy}
+                loading={approveMut.isPending}
+              >
+                Approve + build
+              </Button>
+            </Group>
+          </Stack>
+        )}
+      </Stack>
+    </Modal>
+  );
+}
+
+// ---------------------------------------------------------------------------
+// MOD REVIEW SANDBOX (#2831) — run a PENDING version in a temporary, mod-gated
+// preview before approving. The Preview button starts a review build via
+// blocks.previewRequest, then polls blocks.getReviewStatus (preview-building →
+// deploying → live | failed) and, once live, embeds the review host in an
+// iframe + offers a deep-link. The whole feature is dark behind the mod-only
+// `app-blocks-review-sandbox` flag — when it's off, previewRequest throws
+// UNAUTHORIZED and the panel surfaces a "not enabled" message instead.
+// ---------------------------------------------------------------------------
+
+function ReviewPreviewPanel({
+  publishRequestId,
+  slug,
+}: {
+  publishRequestId: string;
+  slug: string;
+}) {
+  const features = useFeatureFlags();
+  const utils = trpc.useUtils();
+
+  // Poll the review status. Enabled on mount (not gated on a client "started"
+  // flag) so a preview already live on the row is picked up after a page reload
+  // / modal re-open — getReviewStatus returns the PERSISTED deploy_state, so we
+  // derive the button + live iframe from the server, not from ephemeral React
+  // state (which reset to "Start preview" on reload). building/deploying poll
+  // fast, live polls slow, none/failed do a single fetch then stop (see
+  // refetchInterval).
+  const statusQuery = trpc.blocks.getReviewStatus.useQuery(
+    { publishRequestId },
+    {
+      enabled: !!features?.appBlocks,
+      retry: false,
+      refetchInterval: (query) => {
+        // react-query v5: the callback receives the Query; the data is at
+        // query.state.data (matches the my-submissions.tsx idiom).
+        const s = query.state.data?.state;
+        if (s === 'preview-building' || s === 'preview-deploying') return 3000;
+        // Keep a SLOW poll alive while the preview is live — it detects
+        // approve/reject/teardown state changes. getReviewStatus mints a fresh
+        // `?mr=<token>` previewUrl (120s TTL) on every poll, but the IFRAME src is
+        // DECOUPLED from the poll via pickReviewIframeSrc (see `stableIframeSrc`
+        // below): the iframe keeps its src until the embedded token nears expiry,
+        // then swaps ONCE — so the live preview doesn't hard-reload every minute
+        // (which would wipe in-progress interaction). A 60s poll < 120s TTL still
+        // guarantees a fresh token is always available when a swap is due.
+        if (s === 'preview-live') return 60000;
+        return false; // failed or none → stop polling
+      },
+    }
+  );
+
+  const previewMut = trpc.blocks.previewRequest.useMutation({
+    onSuccess: async () => {
+      showSuccessNotification({ message: `Review build started for ${slug}.` });
+      await utils.blocks.getReviewStatus.invalidate({ publishRequestId });
+    },
+    onError: (e) => {
+      showErrorNotification({ title: 'Could not start preview', error: new Error(e.message) });
+    },
+  });
+
+  const teardownMut = trpc.blocks.teardownPreview.useMutation({
+    onSuccess: async () => {
+      showSuccessNotification({ message: `Review preview torn down for ${slug}.` });
+      // Clears the DB preview state → getReviewStatus returns state:null and the
+      // panel reverts to "Start preview". Also refresh the global panel's count.
+      await Promise.all([
+        utils.blocks.getReviewStatus.invalidate({ publishRequestId }),
+        utils.blocks.listActivePreviews.invalidate(),
+      ]);
+    },
+    onError: (e) => {
+      showErrorNotification({ title: 'Could not tear down preview', error: new Error(e.message) });
+    },
+  });
+
+  const state = statusQuery.data?.state ?? null;
+  const detail = statusQuery.data?.detail;
+  // Prefer the FRESH, mod-bound, short-TTL tokened URL (`?mr=<token>`) the server
+  // mints on every poll when the preview is live — that token is the cross-origin
+  // access bridge the `*.civit.ai` mod-gate forwardAuth verifies on the iframe's
+  // entry document request. Read from the latest query data each render so the
+  // iframe never mounts with an expired token. Fall back to the bare host URL
+  // (e.g. while building) only for display.
+  const url = statusQuery.data?.previewUrl ?? detail?.url;
+  const inProgress = state === 'preview-building' || state === 'preview-deploying';
+  const isLive = state === 'preview-live';
+  const isFailed = state === 'preview-failed';
+
+  // Stabilize the IFRAME src so it does NOT change on every poll. getReviewStatus
+  // mints a fresh `?mr=<token>` previewUrl every 60s poll; binding the iframe
+  // straight to `url` would hard-reload it each minute, wiping in-progress
+  // interaction in the previewed block. pickReviewIframeSrc keeps the embedded
+  // src until its token nears expiry (TTL 120s), then swaps once. The Open-host
+  // button + the `url` display still use the freshest token. Held in a ref so the
+  // chosen src survives re-renders; recomputed each render via the pure helper.
+  const iframeSrcRef = useRef<string | undefined>(undefined);
+  const stableIframeSrc = isLive
+    ? pickReviewIframeSrc(iframeSrcRef.current, url, Date.now())
+    : undefined;
+  // Persist the choice so the next render compares against the embedded token,
+  // not the latest poll's. Clear when the preview leaves the live state.
+  iframeSrcRef.current = stableIframeSrc || undefined;
+
+  return (
+    <Stack gap={4}>
+      <Group gap={6}>
+        <IconWindow size={14} />
+        <Text size="sm" fw={600}>
+          Review preview
+        </Text>
+        {state && (
+          <Badge
+            size="sm"
+            variant="light"
+            color={isLive ? 'green' : isFailed ? 'red' : 'blue'}
+          >
+            {state.replace('preview-', '')}
+          </Badge>
+        )}
+      </Group>
+      <Text size="xs" c="dimmed">
+        Run this pending version in a temporary, mod-only preview before approving.
+        Torn down automatically when you approve or reject.
+      </Text>
+
+      <Group gap="xs">
+        <Button
+          size="xs"
+          variant="light"
+          leftSection={<IconWindow size={14} />}
+          loading={previewMut.isPending}
+          disabled={previewMut.isPending || inProgress}
+          onClick={() => previewMut.mutate({ publishRequestId })}
+        >
+          {state ? 'Rebuild preview' : 'Start preview'}
+        </Button>
+        {isLive && url && (
+          <Button
+            size="xs"
+            variant="default"
+            component="a"
+            href={url}
+            target="_blank"
+            rel="noopener"
+            rightSection={<IconExternalLink size={12} />}
+          >
+            Open review host
+          </Button>
+        )}
+        {state && (
+          // Shown for every non-null preview state INCLUDING preview-failed:
+          // teardownPreview clears any deploy_state that starts with `preview-`
+          // back to null, so a FAILED preview can be dismissed to "Start preview"
+          // (not just rebuilt). Label reflects that a failed row has nothing live
+          // to tear down — it's a dismiss.
+          <Button
+            size="xs"
+            variant="light"
+            color="red"
+            leftSection={<IconX size={12} />}
+            loading={teardownMut.isPending}
+            disabled={teardownMut.isPending}
+            onClick={() => teardownMut.mutate({ publishRequestId })}
+          >
+            {isFailed ? 'Dismiss failed preview' : 'Tear down preview'}
+          </Button>
+        )}
+      </Group>
+
+      {inProgress && (
+        <Text size="xs" c="dimmed">
+          Building + deploying the review preview…
+          {detail?.sha ? ` (sha ${detail.sha.slice(0, 12)})` : ''}
+        </Text>
+      )}
+      {isFailed && (
+        <Alert color="red" variant="light" icon={<IconX size={14} />}>
+          {detail?.error ?? 'Review preview failed.'}
+        </Alert>
+      )}
+      {isLive && stableIframeSrc && (
+        <iframe
+          title={`Review preview for ${slug}`}
+          src={stableIframeSrc}
+          // no-referrer: the `?mr=<token>` entry-token URL must NOT leak via the
+          // `Referer` header to assets the previewed (untrusted) block loads.
+          // (We deliberately do NOT bind the token to a publishRequestId: that
+          // binding isn't statelessly verifiable at the mod-gate without a DB
+          // lookup, which the stateless forwardAuth gate intentionally avoids.
+          // Host + mod + short TTL + no-referrer is the chosen containment.)
+          referrerPolicy="no-referrer"
+          style={{
+            width: '100%',
+            height: 420,
+            border: '1px solid var(--mantine-color-default-border)',
+            borderRadius: 6,
+          }}
+          sandbox="allow-scripts allow-same-origin allow-forms allow-popups"
+        />
+      )}
+      {statusQuery.error && (
+        <Text size="xs" c="dimmed">
+          {statusQuery.error.message}
+        </Text>
+      )}
+    </Stack>
+  );
+}
+
+// ---------------------------------------------------------------------------
+// F-E E5 screenshot review panel — surfaces the publisher-supplied screenshots
+// auto-discovered from the submitted bundle so the mod reviews them before
+// approval (publisher images = abuse vector). Renders <img> from base64 data
+// URLs returned by the mod-only blocks.getPublishRequestScreenshots query (the
+// pending app has no public screenshot URL yet — it isn't approved).
+// ---------------------------------------------------------------------------
+
+function ScreenshotsReviewPanel({ publishRequestId }: { publishRequestId: string }) {
+  const features = useFeatureFlags();
+  const { data, isLoading, error } = trpc.blocks.getPublishRequestScreenshots.useQuery(
+    { publishRequestId },
+    { enabled: !!features?.appBlocks, retry: false }
+  );
+  const items = data?.items ?? [];
+
+  return (
+    <Stack gap={4}>
+      <Group gap={6}>
+        <IconLayoutGrid size={14} />
+        <Text size="sm" fw={600}>
+          Screenshots
+        </Text>
+        {!isLoading && !error && (
+          <Badge size="sm" variant="light" color={items.length > 0 ? 'blue' : 'gray'}>
+            {items.length}
+          </Badge>
+        )}
+      </Group>
+      {isLoading ? (
+        <Text size="xs" c="dimmed">
+          Loading screenshots from the submitted bundle…
+        </Text>
+      ) : error ? (
+        <Text size="xs" c="red">
+          Could not load screenshots: {error.message}
+        </Text>
+      ) : items.length === 0 ? (
+        <Text size="xs" c="dimmed">
+          This bundle includes no `screenshots/` directory.
+        </Text>
+      ) : (
+        <>
+          <Text size="xs" c="dimmed">
+            Publisher-supplied images from the bundle — review before approving.
+          </Text>
+          <SimpleGrid cols={{ base: 1, sm: 2 }} spacing="sm">
+            {items.map((shot) => (
+              <Card key={shot.index} withBorder padding={0} radius="md" style={{ overflow: 'hidden' }}>
+                {/* eslint-disable-next-line @next/next/no-img-element */}
+                <img
+                  src={shot.dataUrl}
+                  alt={`Screenshot ${shot.index + 1}`}
+                  loading="lazy"
+                  style={{ width: '100%', height: 'auto', display: 'block' }}
+                />
+              </Card>
+            ))}
+          </SimpleGrid>
+        </>
+      )}
+    </Stack>
+  );
+}
+
+// ---------------------------------------------------------------------------
+// F-E E4 curation panel — mod controls to set the marketplace category +
+// featured/order on an approved app_block (calls blocks.setMarketplaceMeta).
+// Mod-only (the trPC procedures are moderatorProcedure-gated; the whole review
+// page already requires isModerator).
+// ---------------------------------------------------------------------------
+
+const CATEGORY_SELECT_DATA = MARKETPLACE_CATEGORIES.map((c) => ({
+  value: c,
+  label: MARKETPLACE_CATEGORY_LABELS[c],
+}));
+
+function CurationPanel({ appBlockId }: { appBlockId: string }) {
+  const features = useFeatureFlags();
+  const utils = trpc.useUtils();
+  const metaQuery = trpc.blocks.getMarketplaceMeta.useQuery(
+    { appBlockId },
+    { enabled: !!features?.appBlocks }
+  );
+
+  // Local form state, seeded from the fetched meta. `dirty` becomes true once
+  // the mod edits a field so Save is only enabled when there's a change.
+  const [category, setCategory] = useState<MarketplaceCategory | null>(null);
+  const [featured, setFeatured] = useState(false);
+  const [featuredOrder, setFeaturedOrder] = useState<number | null>(null);
+  const [seeded, setSeeded] = useState(false);
+
+  const meta = metaQuery.data;
+  // Seed once when the query first resolves (and re-seed if a different app's
+  // meta arrives — keyed on appBlockId via the parent remount, but guard here
+  // too in case the panel is reused).
+  if (meta && !seeded) {
+    setCategory(
+      meta.category && (MARKETPLACE_CATEGORIES as readonly string[]).includes(meta.category)
+        ? (meta.category as MarketplaceCategory)
+        : null
+    );
+    setFeatured(meta.featured);
+    setFeaturedOrder(meta.featuredOrder);
+    setSeeded(true);
+  }
+
+  const saveMut = trpc.blocks.setMarketplaceMeta.useMutation({
+    onSuccess: async () => {
+      showSuccessNotification({ message: 'Marketplace metadata saved.' });
+      await utils.blocks.getMarketplaceMeta.invalidate({ appBlockId });
+      await utils.blocks.getFeaturedBlocks.invalidate();
+      await utils.blocks.listAvailable.invalidate();
+    },
+    onError: (e) => {
+      showErrorNotification({ title: 'Save failed', error: new Error(e.message) });
+    },
+  });
+
+  const isApproved = meta?.status === 'approved';
+  const busy = saveMut.isPending;
+
+  const dirty =
+    !!meta &&
+    ((category ?? null) !== (meta.category ?? null) ||
+      featured !== meta.featured ||
+      (featuredOrder ?? null) !== (meta.featuredOrder ?? null));
+
+  return (
+    <Card withBorder p="sm">
+      <Stack gap="sm">
+        <Group gap={6}>
+          <IconLayoutGrid size={14} />
+          <Text size="sm" fw={600}>
+            Marketplace curation
+          </Text>
+        </Group>
+
+        {metaQuery.isLoading ? (
+          <Text size="xs" c="dimmed">
+            Loading…
+          </Text>
+        ) : metaQuery.isError ? (
+          <Alert color="red" icon={<IconAlertTriangle size={16} />}>
+            {metaQuery.error.message}
+          </Alert>
+        ) : (
+          <>
+            {!isApproved && (
+              <Alert color="yellow" variant="light" icon={<IconAlertTriangle size={16} />}>
+                This app is{' '}
+                <Code>{meta?.status ?? 'unknown'}</Code> — only an approved app can be
+                featured. Category/order can still be set.
+              </Alert>
+            )}
+            <Select
+              label="Category"
+              data={CATEGORY_SELECT_DATA}
+              value={category}
+              onChange={(v) => setCategory((v as MarketplaceCategory) || null)}
+              placeholder="No category"
+              clearable
+              disabled={busy}
+              w={240}
+            />
+            <Switch
+              label="Featured (staff pick rail)"
+              checked={featured}
+              onChange={(e) => setFeatured(e.currentTarget.checked)}
+              disabled={busy || !isApproved}
+            />
+            <NumberInput
+              label="Featured order (lower = earlier)"
+              value={featuredOrder ?? ''}
+              onChange={(v) =>
+                setFeaturedOrder(typeof v === 'number' ? v : v === '' ? null : Number(v))
+              }
+              min={0}
+              max={100000}
+              allowDecimal={false}
+              placeholder="unset"
+              disabled={busy}
+              w={240}
+            />
+            <Group justify="flex-end">
+              <Button
+                size="xs"
+                leftSection={<IconCheck size={14} />}
+                disabled={!dirty || busy}
+                loading={busy}
+                onClick={() =>
+                  saveMut.mutate({
+                    appBlockId,
+                    category: category ?? null,
+                    featured,
+                    featuredOrder: featuredOrder ?? null,
+                  })
+                }
+              >
+                Save curation
+              </Button>
+            </Group>
+          </>
+        )}
+      </Stack>
+    </Card>
+  );
+}
+
+// ---------------------------------------------------------------------------
+// Line-level code diff — expands the file-level summary with the actual unified
+// line diff per changed/added text file. Lazy: the query only fires once the mod
+// toggles "Show code diff" on, so the modal default stays light. Bounded
+// server-side (text-only, byte/line/file caps); elided files render a "view in
+// Forgejo" fallback rather than inlining unbounded content.
+//
+// The presentational panels (FileListPreview / FileDiffEntry / DiffHunkView /
+// ManifestDiffPreview) + the FileLineDiff type live in the server-free
+// `~/components/Apps/reviewDiffPanels` module so they can be unit tested in
+// browser mode without importing this page's tRPC server graph.
+// ---------------------------------------------------------------------------
+
+function CodeDiffPanel({
+  publishRequestId,
+  forgejoUrl,
+}: {
+  publishRequestId: string;
+  forgejoUrl: string;
+}) {
+  const features = useFeatureFlags();
+  const [show, setShow] = useState(false);
+
+  const { data, isLoading, error } = trpc.blocks.getPublishRequestDiff.useQuery(
+    { publishRequestId },
+    // Only fetch once toggled on (lazy). retry:false keeps a failed fetch from
+    // hammering MinIO/Forgejo for a request with no diffable artifact.
+    { enabled: !!features?.appBlocks && show, retry: false }
+  );
+
+  const files = (data?.files ?? []) as FileLineDiff[];
+
+  return (
+    <Stack gap={4}>
+      <Group gap={8}>
+        <Switch
+          size="sm"
+          label="Show code diff"
+          checked={show}
+          onChange={(e) => setShow(e.currentTarget.checked)}
+        />
+        {show && !isLoading && !error && (
+          <Text size="xs" c="dimmed">
+            {files.length} file{files.length === 1 ? '' : 's'} changed
+            {data?.truncated ? ' (some elided — view in Forgejo)' : ''}
+          </Text>
+        )}
+      </Group>
+
+      {show &&
+        (isLoading ? (
+          <Text size="xs" c="dimmed">
+            Computing line diff from the submitted bundle…
+          </Text>
+        ) : error ? (
+          <Text size="xs" c="red">
+            Could not load code diff: {error.message}
+          </Text>
+        ) : files.length === 0 ? (
+          <Text size="xs" c="dimmed">
+            No textual file changes to show.
+          </Text>
+        ) : (
+          <Stack gap={6}>
+            {files.map((f) => (
+              <FileDiffEntry key={f.path} file={f} forgejoUrl={forgejoUrl} />
+            ))}
+          </Stack>
+        ))}
+    </Stack>
+  );
+}
+
+// ---------------------------------------------------------------------------
+// Structured manifest renderer (replaces the raw JSON dump)
+// ---------------------------------------------------------------------------
+
+/** Manifest top-level keys this renderer handles inline. Anything else falls
+ * into the "Other fields" accordion as raw JSON so reviewers can still see
+ * unexpected payloads. */
+const HANDLED_MANIFEST_KEYS = new Set([
+  '$schema',
+  'appId',
+  'blockId',
+  'version',
+  'name',
+  'description',
+  'type',
+  'minApiVersion',
+  'contentRating',
+  'renderMode',
+  'trustTier',
+  'scopes',
+  'targets',
+  'iframe',
+  'settings',
+]);
+
+function ManifestView({ manifest }: { manifest: Record<string, unknown> }) {
+  const otherKeys = useMemo(
+    () =>
+      Object.keys(manifest)
+        .filter((k) => !HANDLED_MANIFEST_KEYS.has(k))
+        .sort(),
+    [manifest]
+  );
+
+  return (
+    <Stack gap="sm">
+      <ManifestIdentity manifest={manifest} />
+      <ManifestScopes manifest={manifest} />
+      <ManifestTargets manifest={manifest} />
+      <ManifestIframe manifest={manifest} />
+      <ManifestSettings manifest={manifest} />
+      {otherKeys.length > 0 && (
+        <Accordion variant="contained" multiple={false}>
+          <Accordion.Item value="other">
+            <Accordion.Control>
+              <Text size="sm" fw={500}>
+                Other manifest fields ({otherKeys.length})
+              </Text>
+            </Accordion.Control>
+            <Accordion.Panel>
+              <ScrollArea h={200}>
+                <Code block style={{ fontSize: 11, padding: 8 }}>
+                  {JSON.stringify(
+                    Object.fromEntries(otherKeys.map((k) => [k, manifest[k]])),
+                    null,
+                    2
+                  )}
+                </Code>
+              </ScrollArea>
+            </Accordion.Panel>
+          </Accordion.Item>
+        </Accordion>
+      )}
+    </Stack>
+  );
+}
+
+function ManifestIdentity({ manifest }: { manifest: Record<string, unknown> }) {
+  const name = typeof manifest.name === 'string' ? manifest.name : null;
+  const description =
+    typeof manifest.description === 'string' ? manifest.description : null;
+  const blockId = typeof manifest.blockId === 'string' ? manifest.blockId : null;
+  const version = typeof manifest.version === 'string' ? manifest.version : null;
+  const contentRating =
+    typeof manifest.contentRating === 'string' ? manifest.contentRating : null;
+  const trustTier = typeof manifest.trustTier === 'string' ? manifest.trustTier : null;
+  const renderMode = typeof manifest.renderMode === 'string' ? manifest.renderMode : null;
+
+  return (
+    <Card withBorder p="sm">
+      <Stack gap="xs">
+        <Group justify="space-between" align="flex-start">
+          <Stack gap={2}>
+            {name && (
+              <Text size="md" fw={600}>
+                {name}
+              </Text>
+            )}
+            <Group gap={6}>
+              {blockId && <Code>{blockId}</Code>}
+              {version && (
+                <Badge color="gray" variant="light">
+                  v{version}
+                </Badge>
+              )}
+            </Group>
+          </Stack>
+          <Group gap={6}>
+            {contentRating && (
+              <Tooltip label="Content rating">
+                <Badge color={ratingColor(contentRating)} variant="filled">
+                  {contentRating}
+                </Badge>
+              </Tooltip>
+            )}
+            {trustTier && (
+              <Tooltip label="Trust tier">
+                <Badge color={trustColor(trustTier)} variant="light">
+                  {trustTier}
+                </Badge>
+              </Tooltip>
+            )}
+            {renderMode && (
+              <Tooltip label="Render mode">
+                <Badge color="gray" variant="outline">
+                  {renderMode}
+                </Badge>
+              </Tooltip>
+            )}
+          </Group>
+        </Group>
+        {description && (
+          <Text size="sm" c="dimmed" style={{ whiteSpace: 'pre-wrap' }}>
+            {description}
+          </Text>
+        )}
+      </Stack>
+    </Card>
+  );
+}
+
+function ratingColor(rating: string): string {
+  switch (rating.toLowerCase()) {
+    case 'g':
+      return 'green';
+    case 'pg':
+      return 'lime';
+    case 'pg13':
+      return 'yellow';
+    case 'r':
+      return 'orange';
+    case 'x':
+    case 'xxx':
+      return 'red';
+    default:
+      return 'gray';
+  }
+}
+
+function trustColor(tier: string): string {
+  switch (tier.toLowerCase()) {
+    case 'internal':
+      return 'blue';
+    case 'verified':
+      return 'green';
+    case 'unverified':
+      return 'orange';
+    default:
+      return 'gray';
+  }
+}
+
+function ManifestScopes({ manifest }: { manifest: Record<string, unknown> }) {
+  const scopes = Array.isArray(manifest.scopes)
+    ? (manifest.scopes as unknown[]).filter((s): s is string => typeof s === 'string')
+    : [];
+  return (
+    <Card withBorder p="sm">
+      <Stack gap="xs">
+        <Group gap={6}>
+          <IconKey size={14} />
+          <Text size="sm" fw={600}>
+            JWT scopes ({scopes.length})
+          </Text>
+        </Group>
+        {scopes.length === 0 ? (
+          <Text size="xs" c="dimmed" fs="italic">
+            No scopes requested — block can only consume host postMessage
+            data, no scope-gated platform APIs.
+          </Text>
+        ) : (
+          <Stack gap={4}>
+            {scopes.map((s) => {
+              const desc = SCOPE_DESCRIPTIONS[s];
+              const known = !!desc;
+              return (
+                <Group key={s} gap={8} align="flex-start" wrap="nowrap">
+                  <Badge
+                    variant={known ? 'light' : 'outline'}
+                    color={known ? 'blue' : 'red'}
+                    style={{ fontFamily: 'ui-monospace, monospace' }}
+                  >
+                    {s}
+                  </Badge>
+                  <Text size="xs" c={known ? 'dimmed' : 'red'}>
+                    {desc ?? 'Unknown scope — would fail at token issuance.'}
+                  </Text>
+                </Group>
+              );
+            })}
+          </Stack>
+        )}
+      </Stack>
+    </Card>
+  );
+}
+
+function ManifestTargets({ manifest }: { manifest: Record<string, unknown> }) {
+  const targets = Array.isArray(manifest.targets)
+    ? (manifest.targets as Array<Record<string, unknown>>)
+    : [];
+  if (targets.length === 0) return null;
+  return (
+    <Card withBorder p="sm">
+      <Stack gap="xs">
+        <Group gap={6}>
+          <IconLayoutGrid size={14} />
+          <Text size="sm" fw={600}>
+            Slot targets ({targets.length})
+          </Text>
+        </Group>
+        <Stack gap={6}>
+          {targets.map((t, i) => {
+            const slotId = typeof t.slotId === 'string' ? t.slotId : '?';
+            const priority = typeof t.priority === 'number' ? t.priority : null;
+            const requiredContext = Array.isArray(t.requiredContext)
+              ? (t.requiredContext as unknown[]).filter(
+                  (c): c is string => typeof c === 'string'
+                )
+              : [];
+            const slotDesc = SLOT_DESCRIPTIONS[slotId];
+            return (
+              <Stack key={i} gap={2}>
+                <Group gap={6}>
+                  <Code>{slotId}</Code>
+                  {priority !== null && (
+                    <Badge size="xs" color="gray" variant="light">
+                      priority {priority}
+                    </Badge>
+                  )}
+                </Group>
+                {slotDesc && (
+                  <Text size="xs" c="dimmed" pl={4}>
+                    {slotDesc}
+                  </Text>
+                )}
+                {requiredContext.length > 0 && (
+                  <Group gap={4} pl={4}>
+                    <Text size="xs" c="dimmed">
+                      requires:
+                    </Text>
+                    {requiredContext.map((c) => (
+                      <Code key={c} style={{ fontSize: 10 }}>
+                        {c}
+                      </Code>
+                    ))}
+                  </Group>
+                )}
+              </Stack>
+            );
+          })}
+        </Stack>
+      </Stack>
+    </Card>
+  );
+}
+
+function ManifestIframe({ manifest }: { manifest: Record<string, unknown> }) {
+  const iframe =
+    manifest.iframe && typeof manifest.iframe === 'object'
+      ? (manifest.iframe as Record<string, unknown>)
+      : null;
+  if (!iframe) return null;
+  const src = typeof iframe.src === 'string' ? iframe.src : null;
+  const sandbox = typeof iframe.sandbox === 'string' ? iframe.sandbox : null;
+  const minHeight = typeof iframe.minHeight === 'number' ? iframe.minHeight : null;
+  const maxHeight = typeof iframe.maxHeight === 'number' ? iframe.maxHeight : null;
+  const resizable = typeof iframe.resizable === 'boolean' ? iframe.resizable : null;
+
+  const sandboxFlags = sandbox ? sandbox.split(/\s+/).filter(Boolean) : [];
+
+  return (
+    <Card withBorder p="sm">
+      <Stack gap="xs">
+        <Group gap={6}>
+          <IconWindow size={14} />
+          <Text size="sm" fw={600}>
+            Iframe
+          </Text>
+        </Group>
+        {src && (
+          <Group gap={6} align="baseline">
+            <Text size="xs" c="dimmed" style={{ minWidth: 60 }}>
+              src
+            </Text>
+            <a href={src} target="_blank" rel="noopener" style={{ fontSize: 12 }}>
+              {src}
+            </a>
+          </Group>
+        )}
+        {sandboxFlags.length > 0 && (
+          <Group gap={6} align="baseline">
+            <Text size="xs" c="dimmed" style={{ minWidth: 60 }}>
+              sandbox
+            </Text>
+            <Group gap={4}>
+              {sandboxFlags.map((flag) => {
+                const risky =
+                  flag === 'allow-same-origin' ||
+                  flag === 'allow-top-navigation' ||
+                  flag === 'allow-popups-to-escape-sandbox';
+                return (
+                  <Tooltip
+                    key={flag}
+                    label={
+                      risky
+                        ? 'Higher-risk sandbox flag — review carefully.'
+                        : 'Standard sandbox flag.'
+                    }
+                  >
+                    <Badge
+                      size="xs"
+                      color={risky ? 'orange' : 'gray'}
+                      variant="light"
+                      leftSection={risky ? <IconShieldLock size={10} /> : undefined}
+                    >
+                      {flag}
+                    </Badge>
+                  </Tooltip>
+                );
+              })}
+            </Group>
+          </Group>
+        )}
+        {(minHeight !== null || maxHeight !== null || resizable !== null) && (
+          <Group gap={12}>
+            {minHeight !== null && (
+              <Text size="xs">
+                <Text component="span" size="xs" c="dimmed">
+                  min height:
+                </Text>{' '}
+                {minHeight}px
+              </Text>
+            )}
+            {maxHeight !== null && (
+              <Text size="xs">
+                <Text component="span" size="xs" c="dimmed">
+                  max height:
+                </Text>{' '}
+                {maxHeight}px
+              </Text>
+            )}
+            {resizable !== null && (
+              <Text size="xs">
+                <Text component="span" size="xs" c="dimmed">
+                  resizable:
+                </Text>{' '}
+                {resizable ? 'yes' : 'no'}
+              </Text>
+            )}
+          </Group>
+        )}
+      </Stack>
+    </Card>
+  );
+}
+
+function ManifestSettings({ manifest }: { manifest: Record<string, unknown> }) {
+  const settings =
+    manifest.settings && typeof manifest.settings === 'object'
+      ? (manifest.settings as Record<string, unknown>)
+      : null;
+  const entries = settings ? Object.entries(settings) : [];
+  if (entries.length === 0) return null;
+  return (
+    <Card withBorder p="sm">
+      <Stack gap="xs">
+        <Group gap={6}>
+          <IconAdjustmentsAlt size={14} />
+          <Text size="sm" fw={600}>
+            Settings ({entries.length})
+          </Text>
+        </Group>
+        <Stack gap={8}>
+          {entries.map(([key, raw]) => {
+            const def =
+              raw && typeof raw === 'object' ? (raw as Record<string, unknown>) : {};
+            const type = typeof def.type === 'string' ? def.type : '?';
+            const widget = typeof def.widget === 'string' ? def.widget : null;
+            const label = typeof def.label === 'string' ? def.label : null;
+            const description =
+              typeof def.description === 'string' ? def.description : null;
+            const scope = typeof def.scope === 'string' ? def.scope : null;
+            const defaultVal = def.default;
+            const min = typeof def.min === 'number' ? def.min : null;
+            const max = typeof def.max === 'number' ? def.max : null;
+            const requiresScope =
+              typeof def.requires_scope === 'string' ? def.requires_scope : null;
+            return (
+              <Stack key={key} gap={2}>
+                <Group gap={6} wrap="nowrap" align="baseline">
+                  <Code style={{ fontSize: 11 }}>{key}</Code>
+                  <Badge size="xs" variant="light">
+                    {type}
+                    {widget && widget !== type ? `/${widget}` : ''}
+                  </Badge>
+                  {scope && (
+                    <Badge size="xs" color="gray" variant="outline">
+                      {scope}
+                    </Badge>
+                  )}
+                  {requiresScope && (
+                    <Badge size="xs" color="blue" variant="outline">
+                      needs {requiresScope}
+                    </Badge>
+                  )}
+                </Group>
+                {label && (
+                  <Text size="xs" pl={4}>
+                    {label}
+                  </Text>
+                )}
+                {description && (
+                  <Text size="xs" c="dimmed" pl={4}>
+                    {description}
+                  </Text>
+                )}
+                <Group gap={8} pl={4}>
+                  {defaultVal !== undefined && (
+                    <Text size="xs" c="dimmed">
+                      default:{' '}
+                      <Code style={{ fontSize: 10 }}>{JSON.stringify(defaultVal)}</Code>
+                    </Text>
+                  )}
+                  {(min !== null || max !== null) && (
+                    <Text size="xs" c="dimmed">
+                      range: {min ?? '−∞'} – {max ?? '+∞'}
+                    </Text>
+                  )}
+                </Group>
+              </Stack>
+            );
+          })}
+        </Stack>
+      </Stack>
+    </Card>
+  );
+}
